Refactoring part 1
