+++ memoryHandler.h
@@ -1,0 +1,11 @@
+//
+// Created by Anca on 12/22/2019.
+//
+
+#ifndef FOOD_ORDERING_MEMORYHANDLER_H
+#define FOOD_ORDERING_MEMORYHANDLER_H
+#include "create.h"
+
+void freeData(char line, char lineDrinks, Type * types, int noOfTypes, Drink* drinks, int noOfDrinks);
+
+#endif //FOOD_ORDERING_MEMORYHANDLER_H
